Removed unused tasks.
